add readme

+++ src/main/java/com/fishTrade/repository/ICustomerRepository.java
@@ -3,5 +3,5 @@ package com.fishTrade.repository;
 import com.fishTrade.entity.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ICustomerRepository extends JpaRepository<Customer, Long> {
+public interface ICustomerRepository extends JpaRepository<Customer, Integer> {
 }
